+++ LMeter/ACT/ACTClient.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Net.WebSockets;
 using System.Text;
@@ -7,6 +8,7 @@ using System.Threading.Tasks;
 using Dalamud.Game.Gui;
 using Dalamud.Game.Text;
 using Dalamud.Logging;
+using LMeter.Config;
 using LMeter.Helpers;
 using Newtonsoft.Json;
 
@@ -23,30 +25,34 @@ namespace LMeter.ACT
 
     public class ACTClient : ILMeterDisposable
     {
+        private ACTConfig _config;
         private ClientWebSocket _socket;
         private CancellationTokenSource _cancellationTokenSource;
         private Task? _receiveTask;
         private ACTEvent? _lastEvent;
         private ConnectionStatus _status;
-
-        public ACTClient()
-        {
-            this._socket = new ClientWebSocket();
-            this._cancellationTokenSource = new CancellationTokenSource();
-            this._status = ConnectionStatus.NotConnected;
-        }
+        private List<ACTEvent> _pastEvents;
 
         public static ConnectionStatus Status => Singletons.Get<ACTClient>()._status;
+        public static List<ACTEvent> PastEvents => Singletons.Get<ACTClient>()._pastEvents;
 
-        public static ACTEvent? GetLastEvent()
+        public ACTClient(ACTConfig config)
+        {
+            _config = config;
+            _socket = new ClientWebSocket();
+            _cancellationTokenSource = new CancellationTokenSource();
+            _status = ConnectionStatus.NotConnected;
+            _pastEvents = new List<ACTEvent>();
+        }
+
+        public static ACTEvent? GetEvent(int index)
         {
             ACTClient client = Singletons.Get<ACTClient>();
-            if (client._status != ConnectionStatus.Connected ||
-                client._lastEvent is null)
+            if (index >= 0 && index < client._pastEvents.Count)
             {
-                return null;
+                return client._pastEvents[index];
             }
-
+            
             return client._lastEvent;
         }
 
@@ -62,10 +68,11 @@ namespace LMeter.ACT
             chat.PrintChat(message);
         }
 
-        public static void Clear(bool clearAct)
+        public void Clear()
         {
-            Singletons.Get<ACTClient>()._lastEvent = null;
-            if (clearAct)
+            _lastEvent = null;
+            _pastEvents = new List<ACTEvent>();
+            if (_config.ClearACT)
             {
                 ChatGui chat = Singletons.Get<ChatGui>();
                 XivChatEntry message = new XivChatEntry()
@@ -82,12 +89,12 @@ namespace LMeter.ACT
         {
             ACTClient client = Singletons.Get<ACTClient>();
             client.Reset();
-            client.Start(address);
+            client.Start();
         }
 
-        public void Start(string host)
+        public void Start()
         {
-            if (this._status != ConnectionStatus.NotConnected)
+            if (_status != ConnectionStatus.NotConnected)
             {
                 PluginLog.Error("Cannot start, ACTClient needs to be reset!");
                 return;
@@ -95,11 +102,11 @@ namespace LMeter.ACT
 
             try
             {
-                this._receiveTask = Task.Run(() => this.Connect(host));
+                _receiveTask = Task.Run(() => this.Connect(_config.ACTSocketAddress));
             }
             catch (Exception ex)
             {
-                this._status = ConnectionStatus.ConnectionFailed;
+                _status = ConnectionStatus.ConnectionFailed;
                 this.LogConnectionFailure(ex.ToString());
             }
         }
@@ -108,19 +115,19 @@ namespace LMeter.ACT
         {
             try
             {
-                this._status = ConnectionStatus.Connecting;
-                await this._socket.ConnectAsync(new Uri(host), this._cancellationTokenSource.Token);
+                _status = ConnectionStatus.Connecting;
+                await _socket.ConnectAsync(new Uri(host), _cancellationTokenSource.Token);
 
                 string subscribe = "{\"call\":\"subscribe\",\"events\":[\"CombatData\"]}";
-                await this._socket.SendAsync(
+                await _socket.SendAsync(
                         Encoding.UTF8.GetBytes(subscribe),
                         WebSocketMessageType.Text,
                         endOfMessage: true,
-                        this._cancellationTokenSource.Token);
+                        _cancellationTokenSource.Token);
             }
             catch (Exception ex)
             {
-                this._status = ConnectionStatus.ConnectionFailed;
+                _status = ConnectionStatus.ConnectionFailed;
                 this.LogConnectionFailure(ex.ToString());
                 return;
             }
@@ -128,12 +135,12 @@ namespace LMeter.ACT
             ArraySegment<byte> buffer = new ArraySegment<byte>(new byte[4096]);
             if (buffer.Array is null)
             {
-                this._status = ConnectionStatus.ConnectionFailed;
+                _status = ConnectionStatus.ConnectionFailed;
                 this.LogConnectionFailure("Failed to allocate receive buffer!");
                 return;
             }
 
-            this._status = ConnectionStatus.Connected;
+            _status = ConnectionStatus.Connected;
             PluginLog.Information("Successfully Established ACT Connection");
             try
             {
@@ -144,7 +151,7 @@ namespace LMeter.ACT
                     {
                         do
                         {
-                            result = await this._socket.ReceiveAsync(buffer, this._cancellationTokenSource.Token);
+                            result = await _socket.ReceiveAsync(buffer, _cancellationTokenSource.Token);
                             ms.Write(buffer.Array, buffer.Offset, result.Count);
                         }
                         while (!result.EndOfMessage);
@@ -164,14 +171,25 @@ namespace LMeter.ACT
                             {
                                 try
                                 {
-                                    ACTEvent? actEvent = JsonConvert.DeserializeObject<ACTEvent>(data);
+                                    ACTEvent? newEvent = JsonConvert.DeserializeObject<ACTEvent>(data);
 
-                                    if (actEvent?.Combatants is not null &&
-                                        actEvent.Combatants.Count > 0 &&
-                                        (CharacterState.IsInCombat() || !actEvent.IsEncounterActive()))
+                                    if (newEvent?.Encounter is not null &&
+                                        newEvent?.Combatants is not null &&
+                                        newEvent.Combatants.Count > 0 &&
+                                        (CharacterState.IsInCombat() || !newEvent.IsEncounterActive()))
                                     {
-                                        actEvent.Timestamp = DateTime.UtcNow;
-                                        this._lastEvent = actEvent;
+                                        if (!newEvent.IsEncounterActive())
+                                        {
+                                            _pastEvents.Add(newEvent);
+
+                                            while (_pastEvents.Count > _config.EncounterHistorySize)
+                                            {
+                                                _pastEvents.RemoveAt(0);
+                                            }
+                                        }
+
+                                        newEvent.Timestamp = DateTime.UtcNow;
+                                        _lastEvent = newEvent;
                                     }
                                 }
                                 catch (Exception ex)
@@ -182,7 +200,7 @@ namespace LMeter.ACT
                         }
                     }
                 }
-                while (this._status == ConnectionStatus.Connected);
+                while (_status == ConnectionStatus.Connected);
             }
             catch
             {
@@ -190,7 +208,7 @@ namespace LMeter.ACT
             }
             finally
             {
-                if (this._status != ConnectionStatus.ShuttingDown)
+                if (_status != ConnectionStatus.ShuttingDown)
                 {
                     this.Shutdown();
                 }
@@ -199,42 +217,42 @@ namespace LMeter.ACT
 
         public void Shutdown()
         {
-            this._status = ConnectionStatus.ShuttingDown;
-            this._lastEvent = null;
-            if (this._socket.State == WebSocketState.Open ||
-                this._socket.State == WebSocketState.Connecting)
+            _status = ConnectionStatus.ShuttingDown;
+            _lastEvent = null;
+            if (_socket.State == WebSocketState.Open ||
+                _socket.State == WebSocketState.Connecting)
             {
                 try
                 {
                     // Close the websocket
-                    this._socket.CloseOutputAsync(WebSocketCloseStatus.NormalClosure, string.Empty, CancellationToken.None)
+                    _socket.CloseOutputAsync(WebSocketCloseStatus.NormalClosure, string.Empty, CancellationToken.None)
                                .GetAwaiter()
                                .GetResult();
                 }
                 catch
                 {
                     // If closing the socket failed, force it with the cancellation token.
-                    this._cancellationTokenSource.Cancel();
+                    _cancellationTokenSource.Cancel();
                 }
 
-                if (this._receiveTask is not null)
+                if (_receiveTask is not null)
                 {
-                    this._receiveTask.Wait();
+                    _receiveTask.Wait();
                 }
 
                 PluginLog.Information($"Closed ACT Connection");
             }
 
-            this._socket.Dispose();
-            this._status = ConnectionStatus.NotConnected;
+            _socket.Dispose();
+            _status = ConnectionStatus.NotConnected;
         }
 
         public void Reset()
         {
             this.Shutdown();
-            this._socket = new ClientWebSocket();
-            this._cancellationTokenSource = new CancellationTokenSource();
-            this._status = ConnectionStatus.NotConnected;
+            _socket = new ClientWebSocket();
+            _cancellationTokenSource = new CancellationTokenSource();
+            _status = ConnectionStatus.NotConnected;
         }
 
         private void LogConnectionFailure(string error)